+++ packages/base/DateSelect/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @toptal/picasso-date-select
 
+## 1.0.46
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @toptal/picasso-select@3.0.1
+
 ## 1.0.45
 
 ### Patch Changes