+++ packages/picasso-rich-text-editor/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @toptal/picasso-rich-text-editor
 
+## 15.0.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @toptal/picasso-select@3.0.1
+
+## 15.0.1
+
+### Patch Changes
+
+- Updated dependencies [[`1f9def0`](https://github.com/toptal/picasso/commit/1f9def0a6f588d9e96eab0bb4829c6165c04f830)]:
+  - @toptal/picasso-list@4.0.0
+
 ## 15.0.0
 
 ### Major Changes